backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic>=2.8,<3
pydantic-settings>=2.0.3
python-dotenv==1.0.1
httpx==0.27.0
openai==1.35.7
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.2
stripe==9.6.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pytest==8.2.2
pytest-asyncio==0.23.7

# Dev / linting
ruff==0.5.5

